requirements.txt
amqp==5.3.1
asgiref==3.8.1
backports.entry-points-selectable==1.3.0
billiard==4.2.1
celery==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.0
cycler==0.12.1
dj-database-url==3.0.1
Django==4.2.21
django-cors-headers==4.7.0
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
facenet-pytorch==2.6.0
filelock==3.18.0
fonttools==4.58.5
fsspec==2025.5.1
gunicorn==23.0.0
honcho==2.0.0
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
kombu==5.5.3
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==10.2.0
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scipy==1.13.1
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.162
ultralytics-thop==2.0.14
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
zipp==3.23.0